[feat] Improves the allocation of a list literal.
Previously, the "DEF_LIST" instruction allocates a small list that was incremented as needed by the "PUSH_LIST" instruction. For a large, this would mean a huge number of instructions. Now, all elements are compiled and the list is allocated just once in "DEF_LIST" with the count of elements.

Additionally, the concatenation of lists using the "+" operator was removed for now.

+++ src/vm/falcon_object.h
@@ -121,7 +121,7 @@ ObjUpvalue *falconUpvalue(FalconVM *vm, FalconValue *slot);
 ObjClosure *falconClosure(FalconVM *vm, ObjFunction *function);
 ObjClass *falconClass(FalconVM *vm, ObjString *name);
 ObjInstance *falconInstance(FalconVM *vm, ObjClass *class_);
-ObjList *falconList(FalconVM *vm, int size);
+ObjList *falconList(FalconVM *vm, uint16_t size);
 ObjNative *falconNative(FalconVM *vm, FalconNativeFn function, const char *name);
 
 /**
